pyproject.toml
[tool.poetry]
name = "french-words"
version = "0.1.0"
license = "MIT"
description = "App to learn French, word by word"
authors = ["Joffrey Bienvenu <joffreybvn@gmail.com>"]
homepage = "https://github.com/Joffreybvn/french-words"
repository = "https://github.com/Joffreybvn/french-words"
readme = "README.md"
keywords = ["french", "learning", "words", "app"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "3.11.*"
flask-appbuilder = "^4.3.0"
flask-migrate = "^4.0.4"
pywebview = {git = "https://github.com/Joffreybvn/pywebview.git", branch = "feature/hide-menu-bar-when-no-menu", extras = ["qt"]}
pygobject = "^3.42.2"  # QT binding for: Linux "libgl1-mesa-dev"
pyinstaller = "^5.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
ruff = "^0.0.253"
pyre-check = "^0.9.18"

[build-system]
requires = ["poetry>=1.3.2"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79

[tool.ruff]
select = ["E", "F"]
ignore = []
line-length = 79
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.mccabe]
max-complexity = 10
